requirements.txt
pydantic>=2.7.0
qdrant-client>=1.14.0
llama-index-core>=0.14.0
llama-index-llms-openai>=0.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-vector-stores-qdrant>=0.9.0
openai>=1.0.0
llama_cloud
llama-index-readers-file>=0.4.0
PyPDF2
fastapi